# master/server/docs/swagger.json
{
  "swagger": "2.0",
  "paths": {
...
    "/task/{taskID}/result": {
      "get": {
        "description": "retrieve task result",
        "produces": [
          "application/json"
        ],
        "tags": [
          "taskResult"
        ],
        "operationId": "retrieveTaskResult",
        "parameters": [
          {
            "type": "integer",
            "name": "taskID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/TaskResultResp"
            }
          },
          "500": {
            "schema": {
              "$ref": "#/definitions/ErrResp"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ErrResp": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string",
          "x-go-name": "Error",
          "example": "some error"
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/utils"
    },
...
    "TaskResultResp": {
      "type": "object",
      "properties": {
        "command": {
          "type": "string",
          "x-go-name": "Command",
          "example": "/afl/afl-fuzz -i xx -o xx ./test"
        },
        "id": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "ID",
          "example": 1
        },
        "stats": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Stats",
          "example": {
            "crashes": "1"
          }
        },
        "taskid": {
          "type": "integer",
          "format": "uint64",
          "x-go-name": "TaskID",
          "example": 1
        },
        "timeExecuted": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "TimeExecuted",
          "example": 60
        },
        "updateAt": {
          "type": "integer",
          "format": "int64",
          "x-go-name": "UpdateAt",
          "example": 1579996805
        }
      },
      "x-go-package": "github.com/Ch4r1l3/cFuzz/master/server/controller"
    },
...
  }
}